requirements.txt
# VenkatGPT — Python Dependencies
# Install: pip install -r requirements.txt

# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# AI / LLM
anthropic==0.29.0

# Vector Store
faiss-cpu==1.8.0

# Embeddings
sentence-transformers==3.0.0

# Text Processing / RAG
langchain==0.2.5
langchain-community==0.2.5

# PDF Parsing
PyMuPDF==1.24.5

# HTTP Client (GitHub API)
httpx==0.27.0

# Redis
redis[asyncio]==5.0.4

# Data Validation
pydantic==2.7.4
pydantic-settings==2.3.4

# Security / Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AWS SDK (optional, for S3 index storage)
boto3==1.34.144

# Numpy
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
tenacity==8.3.0

# Logging
structlog==24.2.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0

google-generativeai==0.7.2